AshPaperTrail.ChangeBuilders.FullDiff (ash_paper_trail v0.1.4)

Builds a diff of the changeset that is both fairly easy read and includes a complete representation of the changes mades.

Functions

build_changes(attributes, changeset, result)

Return a map of the changes made with a key for each attribute and a value that is a map representing each change. The structure of map representing the each change comes in multiple: simple/native, embedded, union, and array of embedded and array of unions.

%{

subject: %{ from: "subject", to: "new subject" },
body: { unchanged: "body" }

}