AshPaperTrail.ChangeBuilders.FullDiff.Helpers (ash_paper_trail v0.1.4)

Misc helpers for building a full diff of a changeset.

Functions

attribute_change_map(arg)

Builds a simple change map based on the given values.

attribute_change_map({data_present, data, value_present, value})

attribute_changes(data_map, value_map)

Building a map of attribute changes for the embedded resource

build_index_change(from, to)

dump_value(values, attribute)

embedded?(type)

embedded_change_map(arg)

embedded_union?(type, subtype)

map_get_keys(resource, keys)

union?(type)

union_change_map(arg)

unique_id(simple_value, dumped_value)